shim.js: moved Function.prototype.bind fallback to separate file.
